Move enums into separate file. Add an alias for slice of Transformations
